+++ CHANGELOG.md
@@ -5,8 +5,11 @@
 - New settings to configure the number of retry attempts for email & SMS at a specific timeframe, with customization available on the dashboard.
 - New settings for number of tries for email & sms on a specific time stamp also customizable on the platform
 - New command for breaking glass cases that generate a new pair of credentials for the dashboard access (one time use
-  per generation). It also has white-flag to skip 2FA validation and when it logins this account is immediately
+  per generation). It also has white-flag to skip 2FA validation and when it logs in this account is immediately
   disabled.
+- New Activity logs only of the portal (Users interaction and settings management)
+- New UserProfile details page with information about the profiles downloaded per user, connectivity status, event logs and other important information
+- Minor rework on some pages (Reset password per user on the dashboard is now a dedicated page and the edit has also been modified to comply with these changes)
 
 Please make sure to execute the new migration to update and use the new required configuration settings
 - Run the migrations with: